requirements.txt
pyTelegramBotAPI>=4.14.0
flask>=3.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
